Move PythonChangePackageTask to own file
