+++ pipeline/pipelines/grpc_generation_pipeline.py
@@ -15,7 +15,7 @@
 """Pipelines that run gRPC codegen"""
 
 from pipeline.pipelines import code_generation_pipeline as code_gen
-from pipeline.tasks import protoc_tasks, publish_tasks
+from pipeline.tasks import protoc_tasks, publish_tasks, python_grpc_tasks
 from pipeline.utils import task_utils
 
 
@@ -61,7 +61,7 @@ class _JavaGrpcTaskFactory(GrpcTaskFactoryBase):
 class _PythonGrpcTaskFactory(GrpcTaskFactoryBase):
 
     def _get_grpc_codegen_tasks(self, **kwargs):
-        return [protoc_tasks.PythonChangePackageTask,
+        return [python_grpc_tasks.PythonChangePackageTask,
                 protoc_tasks.ProtoDescGenTask,
                 protoc_tasks.ProtoCodeGenTask,
                 protoc_tasks.GrpcPackageMetadataGenTask]
