+++ pipeline/utils/task_utils.py
@@ -68,3 +68,15 @@ def instantiate_tasks(task_class_list, inject):
             name += '-' + inject['api_name']
         tasks.append(task_class(name, inject=inject))
     return tasks
+
+
+def find_protos(proto_paths):
+    """Searches along `proto_path` for .proto files and returns a generator of
+    paths"""
+    if type(proto_paths) is not list:
+        raise ValueError("proto_paths must be a list")
+    for path in proto_paths:
+        for root, _, files in os.walk(path):
+            for proto in files:
+                if os.path.splitext(proto)[1] == '.proto':
+                    yield os.path.join(root, proto)
